ResolvedContainerAppsResource + ManagedEnvironmentTreeItem
