+++ src/commands/createManagedEnvironment/createManagedEnvironment.ts
@@ -5,15 +5,15 @@
 
 import { IActionContext, ICreateChildImplContext } from "@microsoft/vscode-azext-utils";
 import { ext } from "../../extensionVariables";
-import { ManagedEnvironmentTreeItem } from "../../tree/ManagedEnvironmentTreeItem";
+import { ResolvedContainerAppsResource } from "../../tree/ResolvedContainerAppsResource";
 import { SubscriptionTreeItem } from "../../tree/SubscriptionTreeItem";
 import { IManagedEnvironmentContext } from "./IManagedEnvironmentContext";
 
-export async function createManagedEnvironment(context: IActionContext & Partial<ICreateChildImplContext> & Partial<IManagedEnvironmentContext>, node?: SubscriptionTreeItem): Promise<ManagedEnvironmentTreeItem> {
+export async function createManagedEnvironment(context: IActionContext & Partial<ICreateChildImplContext> & Partial<IManagedEnvironmentContext>, node?: SubscriptionTreeItem): Promise<ResolvedContainerAppsResource> {
     if (!node) {
         node = await ext.tree.showTreeItemPicker<SubscriptionTreeItem>(SubscriptionTreeItem.contextValue, context);
     }
 
-    const keNode: ManagedEnvironmentTreeItem = await node.createChild(context);
+    const keNode: ResolvedContainerAppsResource = await node.createChild(context);
     return keNode;
 }
